+++ src/Mapping/MappedEventSubscriber.php
@@ -14,6 +14,7 @@ use Doctrine\Common\Annotations\PsrCachedReader;
 use Doctrine\Common\Annotations\Reader;
 use Doctrine\Common\EventArgs;
 use Doctrine\Common\EventSubscriber;
+use Doctrine\Deprecations\Deprecation;
 use Doctrine\ODM\MongoDB\DocumentManager;
 use Doctrine\ODM\MongoDB\Mapping\ClassMetadata as DocumentClassMetadata;
 use Doctrine\ORM\EntityManagerInterface;
@@ -25,10 +26,12 @@ use Doctrine\Persistence\ObjectManager;
 use Gedmo\Exception\InvalidArgumentException;
 use Gedmo\Mapping\Driver\AttributeReader;
 use Gedmo\Mapping\Event\AdapterInterface;
+use Gedmo\Mapping\Event\ClockAwareAdapterInterface;
 use Gedmo\ReferenceIntegrity\Mapping\Validator as ReferenceIntegrityValidator;
 use Gedmo\Uploadable\FilenameGenerator\FilenameGeneratorInterface;
 use Gedmo\Uploadable\Mapping\Validator as MappingValidator;
 use Psr\Cache\CacheItemPoolInterface;
+use Psr\Clock\ClockInterface;
 use Symfony\Component\Cache\Adapter\ArrayAdapter;
 
 /**
@@ -93,6 +96,8 @@ abstract class MappedEventSubscriber implements EventSubscriber
      * @var CacheItemPoolInterface|null
      */
     private $cacheItemPool;
+
+    private ?ClockInterface $clock = null;
 
     public function __construct()
     {
@@ -205,9 +210,9 @@ abstract class MappedEventSubscriber implements EventSubscriber
     public function setAnnotationReader($reader)
     {
         if (!$reader instanceof Reader && !$reader instanceof AttributeReader) {
-            trigger_deprecation(
+            Deprecation::trigger(
                 'gedmo/doctrine-extensions',
-                '3.11',
+                'https://github.com/doctrine-extensions/DoctrineExtensions/pull/2558',
                 'Providing an annotation reader which does not implement %s or is not an instance of %s to %s() is deprecated.',
                 Reader::class,
                 AttributeReader::class,
@@ -221,6 +226,11 @@ abstract class MappedEventSubscriber implements EventSubscriber
     final public function setCacheItemPool(CacheItemPoolInterface $cacheItemPool): void
     {
         $this->cacheItemPool = $cacheItemPool;
+    }
+
+    final public function setClock(ClockInterface $clock): void
+    {
+        $this->clock = $clock;
     }
 
     /**
@@ -266,6 +276,10 @@ abstract class MappedEventSubscriber implements EventSubscriber
                     $adapterClass = 'Gedmo\\Mapping\\Event\\Adapter\\'.$m[1];
                 }
                 $this->adapters[$m[1]] = new $adapterClass();
+
+                if ($this->adapters[$m[1]] instanceof ClockAwareAdapterInterface && $this->clock instanceof ClockInterface) {
+                    $this->adapters[$m[1]]->setClock($this->clock);
+                }
             }
             $this->adapters[$m[1]]->setEventArgs($args);