+++ src/SoftDeleteable/SoftDeleteableListener.php
@@ -11,10 +11,12 @@ namespace Gedmo\SoftDeleteable;
 
 use Doctrine\Common\EventArgs;
 use Doctrine\Common\EventManager;
+use Doctrine\Deprecations\Deprecation;
 use Doctrine\ODM\MongoDB\DocumentManager;
 use Doctrine\ODM\MongoDB\UnitOfWork as MongoDBUnitOfWork;
 use Doctrine\ORM\EntityManagerInterface;
 use Doctrine\Persistence\Event\LoadClassMetadataEventArgs;
+use Doctrine\Persistence\Event\ManagerEventArgs;
 use Doctrine\Persistence\Mapping\ClassMetadata;
 use Doctrine\Persistence\ObjectManager;
 use Gedmo\Mapping\MappedEventSubscriber;
@@ -59,6 +61,10 @@ class SoftDeleteableListener extends MappedEventSubscriber
     /**
      * If it's a SoftDeleteable object, update the "deletedAt" field
      * and skip the removal of the object
+     *
+     * @param ManagerEventArgs $args
+     *
+     * @phpstan-param ManagerEventArgs<ObjectManager> $args
      *
      * @return void
      */
@@ -156,12 +162,14 @@ class SoftDeleteableListener extends MappedEventSubscriber
                 || !$parameters[0]->getType() instanceof \ReflectionNamedType
                 || $eventClass !== $parameters[0]->getType()->getName()
             ) {
-                @trigger_error(sprintf(
+                Deprecation::trigger(
+                    'gedmo/doctrine-extensions',
+                    'https://github.com/doctrine-extensions/DoctrineExtensions/pull/2649',
                     'Type-hinting to something different than "%s" in "%s::%s()" is deprecated.',
                     $eventClass,
                     get_class($listener),
                     $reflMethod->getName()
-                ), E_USER_DEPRECATED);
+                );
 
                 return false;
             }