+++ src/Sortable/SortableListener.php
@@ -12,7 +12,10 @@ namespace Gedmo\Sortable;
 use Doctrine\Common\Comparable;
 use Doctrine\Common\EventArgs;
 use Doctrine\Common\Util\ClassUtils;
+use Doctrine\Deprecations\Deprecation;
+use Doctrine\Persistence\Event\LifecycleEventArgs;
 use Doctrine\Persistence\Event\LoadClassMetadataEventArgs;
+use Doctrine\Persistence\Event\ManagerEventArgs;
 use Doctrine\Persistence\Mapping\ClassMetadata;
 use Doctrine\Persistence\ObjectManager;
 use Gedmo\Mapping\MappedEventSubscriber;
@@ -107,6 +110,10 @@ class SortableListener extends MappedEventSubscriber
      * The synchronization of the objects in memory is done in postFlush. This
      * ensures that the positions have been successfully persisted to database.
      *
+     * @param ManagerEventArgs $args
+     *
+     * @phpstan-param ManagerEventArgs<ObjectManager> $args
+     *
      * @return void
      */
     public function onFlush(EventArgs $args)
@@ -151,6 +158,10 @@ class SortableListener extends MappedEventSubscriber
     /**
      * Update maxPositions as needed
      *
+     * @param LifecycleEventArgs $args
+     *
+     * @phpstan-param LifecycleEventArgs<ObjectManager> $args
+     *
      * @return void
      */
     public function prePersist(EventArgs $args)
@@ -175,6 +186,10 @@ class SortableListener extends MappedEventSubscriber
     }
 
     /**
+     * @param LifecycleEventArgs $args
+     *
+     * @phpstan-param LifecycleEventArgs<ObjectManager> $args
+     *
      * @return void
      */
     public function postPersist(EventArgs $args)
@@ -185,6 +200,10 @@ class SortableListener extends MappedEventSubscriber
     }
 
     /**
+     * @param LifecycleEventArgs $args
+     *
+     * @phpstan-param LifecycleEventArgs<ObjectManager> $args
+     *
      * @return void
      */
     public function preUpdate(EventArgs $args)
@@ -195,6 +214,10 @@ class SortableListener extends MappedEventSubscriber
     }
 
     /**
+     * @param LifecycleEventArgs $args
+     *
+     * @phpstan-param LifecycleEventArgs<ObjectManager> $args
+     *
      * @return void
      */
     public function postRemove(EventArgs $args)
@@ -206,6 +229,10 @@ class SortableListener extends MappedEventSubscriber
 
     /**
      * Sync objects in memory
+     *
+     * @param ManagerEventArgs $args
+     *
+     * @phpstan-param ManagerEventArgs<ObjectManager> $args
      *
      * @return void
      */
@@ -271,12 +298,14 @@ class SortableListener extends MappedEventSubscriber
                                     if (is_int($matches)) {
                                         $matches = 0 === $matches;
                                     } else {
-                                        @trigger_error(sprintf(
+                                        Deprecation::trigger(
+                                            'gedmo/doctrine-extensions',
+                                            'https://github.com/doctrine-extensions/DoctrineExtensions/pull/2542',
                                             'Support for "%s" as return type from "%s::compareTo()" is deprecated since'
                                             .' gedmo/doctrine-extensions 3.11 and will be removed in version 4.0. Return "integer" instead.',
                                             gettype($matches),
                                             Comparable::class
-                                        ), E_USER_DEPRECATED);
+                                        );
                                     }
                                 } else {
                                     $matches = $gr == $value;